+++ assets/js/utils/styleUtils.ts
@@ -1,0 +1,4 @@
+export const getFontColor = (colorName: string): string => {
+    const style = getComputedStyle(document.body);
+    return style.getPropertyValue(colorName);
+  };
